+++ lib.commonjs/providers/provider-fallback.js
@@ -93,10 +93,19 @@ function _normalize(value) {
     console.log("Could not serialize", value);
     throw new Error("Hmm...");
 }
-function normalizeResult(value) {
+function normalizeResult(method, value) {
     if ("error" in value) {
         const error = value.error;
-        return { tag: _normalize(error), value: error };
+        let tag;
+        if ((0, index_js_1.isError)(error, "CALL_EXCEPTION")) {
+            tag = _normalize(Object.assign({}, error, {
+                shortMessage: undefined, reason: undefined, info: undefined
+            }));
+        }
+        else {
+            tag = _normalize(error);
+        }
+        return { tag, value: error };
     }
     const result = value.result;
     return { tag: _normalize(result), value: result };
@@ -418,7 +427,7 @@ class FallbackProvider extends abstract_provider_js_1.AbstractProvider {
         const results = [];
         for (const runner of running) {
             if (runner.result != null) {
-                const { tag, value } = normalizeResult(runner.result);
+                const { tag, value } = normalizeResult(req.method, runner.result);
                 results.push({ tag, value, weight: runner.config.weight });
             }
         }
@@ -536,10 +545,10 @@ class FallbackProvider extends abstract_provider_js_1.AbstractProvider {
             const broadcasts = this.#configs.map(async ({ provider, weight }, index) => {
                 try {
                     const result = await provider._perform(req);
-                    results[index] = Object.assign(normalizeResult({ result }), { weight });
+                    results[index] = Object.assign(normalizeResult(req.method, { result }), { weight });
                 }
                 catch (error) {
-                    results[index] = Object.assign(normalizeResult({ error }), { weight });
+                    results[index] = Object.assign(normalizeResult(req.method, { error }), { weight });
                 }
             });
             // As each promise finishes...